% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catdata.R
\name{catdata}
\alias{catdata}
\title{Pull survey responses and theta values from PiLR data}
\usage{
catdata(
  file = NULL,
  day = NULL,
  type = "cat",
  process = TRUE,
  questions = "last"
)
}
\arguments{
\item{file}{Name of path to file to draw data from}

\item{day}{Indicate a particular day to draw data from YYYY-MM-DD}

\item{type}{Select between \code{"cat", "demographics", "weekly"}. Currently
only works with \code{"cat"}}

\item{process}{Logical. If \code{TRUE}, re-formats PiLR data. Recommended to
leave as \code{TRUE}.}

\item{questions}{Determines which set of responses to choose from. Not sure
exactly how to edit.}
}
\value{
A list containing theta values along with standard errors, as well as
table of participant items and responses.
}
\description{
Used for pulling patient data and organizing for therapist output.
}
\examples{
1 + 1
}
