% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classperf.R
\name{classperf}
\alias{classperf}
\title{Classperf}
\usage{
classperf(pred, obs, metric = "sens", p = levels(obs)[2])
}
\arguments{
\item{pred}{TBA}

\item{obs}{TBA}

\item{metric}{TBA}

\item{p}{TBA}
}
\value{
TBA
}
\description{
Classperf
}
\examples{
1 + 1
}
