% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMod.R
\name{coefs}
\alias{coefs}
\title{Return coefficient values from \code{trainFit} model}
\usage{
coefs(fit, inds = FALSE, gamma = 0.5)
}
\arguments{
\item{fit}{Output of \code{\link{trainFit}}}

\item{inds}{TBD}

\item{gamma}{Hyperparameter gamma value for EBIC.}
}
\value{
Coefficient values associated with final model
}
\description{
Description
}
\examples{
1 + 1
}
