% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilrdata.R
\name{createReport}
\alias{createReport}
\title{Create patient report}
\usage{
createReport(participant, id = 0, output = "./")
}
\arguments{
\item{participant}{Raw PiLR data or path to .csv file. If path to file is
provided, this must be an absolute path not a relative path.}

\item{id}{ID number of participant}

\item{output}{Choose the directory of where to output the report. Defaults to
the current working directory.}
}
\value{
Patient report
}
\description{
Description
}
\examples{
1 + 1
}
