% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainFit.R
\name{ctrl}
\alias{ctrl}
\title{Create control parameters for \code{trainFit}}
\usage{
ctrl(
  type = 2,
  method = "cv",
  n = 10,
  cvreps = 3,
  sumfun = "default",
  res = "final",
  preds = "all",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{type}{TBD}

\item{method}{TBD}

\item{n}{TBD}

\item{cvreps}{TBD}

\item{sumfun}{TBD}

\item{res}{TBD}

\item{preds}{TBD}

\item{verbose}{TBD}

\item{...}{Additional arguments}
}
\value{
TBD
}
\description{
Description
}
\examples{
ctrl()
}
