% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{cutoff}
\alias{cutoff}
\title{Extract cutoff mean and SD across CV folds}
\usage{
cutoff(fit, lambda = "best")
}
\arguments{
\item{fit}{output from trainFit}

\item{lambda}{set to "best" to use optimal lambda}
}
\value{
PRC cutoff mean and SD taken across folds for specific lambda value
}
\description{
Extract cutoff mean and SD across CV folds
}
\examples{
1 + 1
}
