% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMod.R
\name{getMod}
\alias{getMod}
\title{Get the final model(s) from \code{trainFit}}
\usage{
getMod(fit, metric = "PRC", row = NULL, crit = "AIC")
}
\arguments{
\item{fit}{Output of \code{\link{trainFit}}}

\item{metric}{Character string for the performance metric (LIST AVAILABLE)}

\item{row}{Only relevant if there are multiple models that maximize the
performance metric. Allows the user to choose which model (based on the row
number) to return information for.}

\item{crit}{Criterion for choosing the final model (unsure)}
}
\value{
TBD
}
\description{
Description
}
\examples{
1 + 1
}
