% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glint.R
\name{glint}
\alias{glint}
\title{Reshaping glinternet output using criterion-based selection}
\usage{
glint(
  x,
  y = 1,
  method = "AIC",
  m = "zzzall",
  type = "default",
  gamma = 0.5,
  nlam = 50,
  useSE = TRUE,
  nfolds = 10,
  allCoef = FALSE,
  fit = NULL,
  lambda = NULL,
  intOnly = FALSE,
  lamNum = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Reshaping glinternet output using criterion-based selection
}
