% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilrdata.R
\name{itemTable}
\alias{itemTable}
\title{Create table of items and responses from PiLR app}
\usage{
itemTable(data, time = "last", ...)
}
\arguments{
\item{data}{Output from \code{\link{pilrdata}}, or raw PiLR data, or path to
PiLR data .csv file.}

\item{time}{Numeric value to indicate which set of responses to use, or set
to \code{"last"} to choose the most recent set of responses.}

\item{...}{Additional arguments for \code{\link{pilrdata}}}
}
\value{
A kable table
}
\description{
Description
}
\examples{
1 + 1
}
