% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LL.R
\name{LL}
\alias{LL}
\title{Log likelihood}
\usage{
LL(fit, y, x = NULL, int = FALSE)
}
\arguments{
\item{fit}{TBA}

\item{y}{tba}

\item{x}{tba}

\item{int}{TBA}
}
\value{
TBA
}
\description{
Log likelihood
}
\examples{
1 + 1
}
