% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilrdata.R
\name{pilrdata}
\alias{pilrdata}
\title{Pull survey responses and theta values from PiLR data}
\usage{
pilrdata(
  file = NULL,
  survey = c("epsi", "idas"),
  day = NULL,
  type = "cat",
  process = TRUE,
  questions = "last"
)
}
\arguments{
\item{file}{PiLR dataframe, or character string indicating the path of where
to find the relevant .csv file.}

\item{survey}{Specify either \code{"epsi"} or \code{"idas"}}

\item{day}{Indicate a particular day to draw data from, using the format
\code{"YYYY-MM-DD"}}

\item{type}{Select between \code{"cat", "demographics", "weekly"}. Currently
only works with \code{"cat"}}

\item{process}{Logical. If \code{TRUE}, re-formats PiLR data and returns
theta values (with standard errors) along with the relevant items and their
associated responses. Setting to \code{FALSE} returns the raw PiLR data for
the CAT surveys.}

\item{questions}{Determines which set of responses to report. Can be a
numerical value or vector to specify which time points to return items and
responses for. Setting \code{questions = "last"} will only return the set
of responses from the most recent time point. Setting \code{questions =
  "all"} will return the set of responses from each time point as a list.}
}
\value{
A list containing theta values along with standard errors, as well as
table of participant items and responses.
}
\description{
Used for pulling patient data and organizing for therapist output.
}
\examples{
1 + 1
}
