% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilrdata.R
\name{plotThetas}
\alias{plotThetas}
\title{Plot thetas from PiLR app data}
\usage{
plotThetas(data, colors = FALSE, zeroLine = TRUE, pointSize = 3, ...)
}
\arguments{
\item{data}{Output from \code{\link{pilrdata}}, or raw PiLR data, or path to
PiLR data .csv file.}

\item{colors}{Logical. Determines whether to use different colors for each
higher-order dimension.}

\item{zeroLine}{Logical. Determines whether to draw a dashed line}

\item{pointSize}{Numerical value for size of points on plot.}

\item{...}{Additional arguments for \code{\link{pilrdata}}}
}
\value{
A plot of theta values and their standard errors for each
higher-order dimension.
}
\description{
Description
}
\examples{
1 + 1
}
