% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRisk.R
\name{predictRisk}
\alias{predictRisk}
\title{Predict risk of relapse from PiLR data}
\usage{
predictRisk(data, ...)
}
\arguments{
\item{data}{PiLR survey data file}

\item{...}{Additional arguments for pilrdata function}
}
\value{
Risk level associated with each timepoint
}
\description{
Predict risk of relapse from PiLR data
}
\examples{
1 + 1
}
