% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relapseRisk.R
\docType{package}
\name{relapseRisk}
\alias{relapseRisk}
\title{relapseRisk: Predict risk of relapse for anorexia nervosa recovery}
\description{
Implements a machine learning algorithm for predicting recovery outcomes
among patients diagnosed with anorexia nervosa. Designed for use with data
from the STAR app in PiLR for a research project funded by the NIH.
}
