% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{sampleCutoffs}
\alias{sampleCutoffs}
\title{Resample data to estimate cutoff sampling distribution}
\usage{
sampleCutoffs(
  formula,
  data,
  niter = 1000,
  seed = NULL,
  prc = TRUE,
  method = c("boot", "up", "down", "rose", "smote")
)
}
\arguments{
\item{formula}{formula for selected model}

\item{data}{training dataset}

\item{niter}{number of iterations to resample}

\item{seed}{numeric value of seed}

\item{prc}{logical, determines whether to find PRC or ROC cutoffs}

\item{method}{resampling method}
}
\value{
vector of cutoff values
}
\description{
Resample data to estimate cutoff sampling distribution
}
\examples{
1 + 1
}
