% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpsim.R
\name{simpsim}
\alias{simpsim}
\title{Simulate class imbalance data}
\usage{
simpsim(
  n = 50,
  p = 3,
  env = FALSE,
  binary = TRUE,
  qt = 0.2,
  levs = c("Class1", "Class2"),
  b = NULL,
  intercept = FALSE,
  mu = 0,
  sigma = NULL
)
}
\arguments{
\item{n}{TBD}

\item{p}{TBD}

\item{env}{TBD}

\item{binary}{TBD}

\item{qt}{TBD}

\item{levs}{TBD}

\item{b}{TBD}

\item{intercept}{TBD}

\item{mu}{TBD}

\item{sigma}{TBD}
}
\value{
TBD
}
\description{
Description
}
\examples{
x <- simpsim()
}
