% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_roc.R
\name{sums}
\alias{sums}
\title{Summarize results across multiple \code{trainFit} modls}
\usage{
sums(res, metric = "default", means = TRUE)
}
\arguments{
\item{res}{List of \code{\link{trainFit}} outputs, or output from
\code{resamples} function from the \code{caret} package.}

\item{metric}{Character string or vector indicating which performance metrics
to summarize}

\item{means}{Logical. Determines whether to return means for the performance
metrics or not.}
}
\value{
TBD
}
\description{
Summarize results across multiple \code{trainFit} modls
}
\examples{
1 + 1
}
