% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_roc.R
\name{test_roc}
\alias{test_roc}
\title{Test ROC results}
\usage{
test_roc(fit, x = NULL, y = NULL, v = TRUE, ci = TRUE)
}
\arguments{
\item{fit}{TBA}

\item{x}{TBA}

\item{y}{TBA}

\item{v}{TBA}

\item{ci}{TBA}
}
\value{
TBA
}
\description{
Description
}
\examples{
1 + 1
}
