% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainFit.R
\name{trainFit}
\alias{trainFit}
\title{Algorithm for iterated resampling and model fitting}
\usage{
trainFit(
  x,
  y,
  k = "default",
  m = "zzzall",
  subsample = "none",
  lams = "default",
  metric = "default",
  pre = NULL,
  grid = NULL,
  model = "glint",
  seed = NULL,
  time = TRUE,
  ...
)
}
\arguments{
\item{x}{TBD}

\item{y}{TBD}

\item{k}{TBD}

\item{m}{TBD}

\item{subsample}{TBD}

\item{lams}{TBD}

\item{metric}{TBD}

\item{pre}{TBD}

\item{grid}{TBD}

\item{model}{TBD}

\item{seed}{TBD}

\item{time}{TBD}

\item{...}{Additional arguments}
}
\value{
TBD
}
\description{
Description
}
\examples{
1 + 1
}
