% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glint.R
\name{glint}
\alias{glint}
\title{Reshaping glinternet output using criterion-based selection}
\usage{
glint(
  x,
  y = 1,
  method = "AIC",
  m = "zzzall",
  type = "default",
  gamma = 0.5,
  nlam = 50,
  useSE = TRUE,
  nfolds = 10,
  allCoef = FALSE,
  fit = NULL,
  lambda = NULL,
  intOnly = FALSE,
  lamNum = FALSE,
  ...
)
}
\arguments{
\item{x}{TBD}

\item{y}{TBD}

\item{method}{TBD}

\item{m}{TBD}

\item{type}{TBD}

\item{gamma}{TBD}

\item{nlam}{TBD}

\item{useSE}{TBD}

\item{nfolds}{TBD}

\item{allCoef}{TBD}

\item{fit}{TBD}

\item{lambda}{TBD}

\item{intOnly}{TBD}

\item{lamNum}{TBD}

\item{...}{Additional arguments}
}
\value{
TBD
}
\description{
Description. May not even be needed as global function.
}
\examples{
1 + 1
}
