% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainFit.R
\name{trainFit}
\alias{trainFit}
\title{Algorithm for iterated resampling and model fitting}
\usage{
trainFit(
  x,
  y,
  k = "default",
  m = "zzzall",
  subsample = "none",
  lams = "default",
  metric = "default",
  pre = NULL,
  grid = NULL,
  model = "glint",
  seed = NULL,
  time = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Algorithm for iterated resampling and model fitting
}
